Basic API exposing record models with no ability to do Image upload
